bugfix 25

+++ src/lib.rs
@@ -68,8 +68,8 @@ impl CatEngine{
         self.window.gl_swap_window();
     }
 
-    pub fn set_fov(&mut self, fov: i16, screen_width: u32, screen_height: u32, near_clamp: u32, far_clamp: u32) {
-        self.renderer.projection = Mat4::perspective_rh_gl(fov as f32, (screen_width / screen_height) as f32, near_clamp as f32, far_clamp as f32);;
+    pub fn set_fov(&mut self, fov: i16, screen_width: u32, screen_height: u32, near_clamp: f32, far_clamp: f32) {
+        self.renderer.projection = Mat4::perspective_rh_gl(fov as f32, (screen_width / screen_height) as f32, near_clamp, far_clamp);;
         self.fov = fov;
     }
